+++ TeamProject_2_Magazzino/src/product/Smartphone.java
@@ -1,9 +1,10 @@
 package product;
 
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 public class Smartphone extends Product{
-    public Smartphone(ProductTypes type, String brand, String model, String description, double displaySize, double storageCap, double purchasePrice, double sellPrice, int itemId) {
+    public Smartphone(ProductTypes type, String brand, String model, String description, double displaySize, double storageCap, BigDecimal purchasePrice, BigDecimal sellPrice, int itemId) {
         super(type, brand, model, description, displaySize, storageCap, purchasePrice, sellPrice, itemId);
     }
